fix: 人工审核增加自定义调整 issue #3007
