+++ src/frontend/devops-pipeline/src/components/AtomPropertyPanel/ManualReviewUserTask.vue
@@ -5,14 +5,10 @@
                 :is="obj.component"
                 v-bind="obj"
                 v-model="element[key]"
-                :params-list="paramsList"
                 :disabled="disabled"
                 :handle-change="handleUpdateElement"
                 :show-content="disabled"
-                :name="key" v-validate.initial="Object.assign({}, obj.rule, { required: obj.required })"
-                @handle-param-type-change="handleParamTypeChange"
-                @handle-update-param="handleUpdateParam"
-                @handle-update-param-id="handleUpdateParamId" />
+                :name="key" v-validate.initial="Object.assign({}, obj.rule, { required: obj.required })" />
         </form-field>
         <accordion show-content show-checkbox>
             <header class="var-header" slot="header">
