+++ src/frontend/devops-pipeline/src/components/AtomPropertyPanel/atomMixin.js
@@ -48,8 +48,6 @@ import GroupIdSelector from '@/components/atomFormField/groupIdSelector'
 import RemoteCurlUrl from '@/components/atomFormField/RemoteCurlUrl'
 import AutoComplete from '@/components/atomFormField/AutoComplete'
 import { urlJoin, rely } from '../../utils/util'
-import { CHECK_PARAM_LIST, CHECK_DEFAULT_PARAM, isMultipleParam } from '@/store/modules/atom/paramsConfig'
-import { deepCopy } from '@/utils/util'
 
 const atomMixin = {
     props: {
@@ -101,14 +99,6 @@ const atomMixin = {
         ]),
         isThirdParty () {
             return this.isThirdPartyContainer(this.container)
-        },
-        paramsList () {
-            return CHECK_PARAM_LIST.map(item => {
-                return {
-                    id: item.id,
-                    name: this.$t(`storeMap.${item.name}`)
-                }
-            })
         }
     },
     methods: {
@@ -161,39 +151,6 @@ const atomMixin = {
                 atom: this.element,
                 [name]: value
             })
-        },
-        handleParamTypeChange (payload) {
-            const params = this.element['params']
-
-            const { value, paramIndex } = payload
-            const newParams = [
-                ...params.slice(0, paramIndex),
-                {
-                    ...deepCopy(CHECK_DEFAULT_PARAM[value]),
-                    key: params[paramIndex].key,
-                    paramIdKey: params[paramIndex].paramIdKey
-                },
-                ...params.slice(paramIndex + 1)
-            ]
-
-            this.handleUpdateElement('params', newParams)
-        },
-        handleUpdateParam (payload) {
-            const { key, value, paramIndex } = payload
-            const param = this.element['params']
-            if (isMultipleParam(param[paramIndex].valueType) && key === 'value') {
-                Object.assign(param[paramIndex], {
-                    [key]: value.join(',')
-                })
-            } else if (param) {
-                Object.assign(param[paramIndex], {
-                    [key]: value
-                })
-            }
-            this.handleUpdateElement('params', param)
-        },
-        handleUpdateParamId (payload) {
-            this.handleUpdateParam(payload)
         },
         handlePath (path = '', getFileName = false) {
             if (path.startsWith('./')) {
